# docs/API_OpenAPI.yaml
openapi: 3.0.3
info:
  title: Trato de Barbados API
  description: |
    API completa para sistema de gestão de barbearias com Next.js App Router.

    ## Autenticação
    Todas as rotas protegidas requerem JWT token do Supabase Auth.

    ## Multi-tenant
    Sistema opera com duas unidades (Trato e BarberBeer) usando RLS.

    ## Rate Limiting
    - 100 requests/min para usuários autenticados
    - 20 requests/min para webhooks
  version: 1.0.0
  contact:
    name: Trato de Barbados
    email: dev@tratodebarbados.com
  license:
    name: MIT
servers:
  - url: https://trato.vercel.app/api
    description: Produção
  - url: http://localhost:3000/api
    description: Desenvolvimento
security:
  - BearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login do usuário
      description: Autentica usuário e retorna session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "barbeiro@trato.com"
                password:
                  type: string
                  minLength: 6
                  example: "senha123"
      responses:
        "200":
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                  session:
                    $ref: "#/components/schemas/Session"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"
...
  /auth/recover:
    post:
      tags: [Auth]
      summary: Recuperação de senha
      description: Envia email de recuperação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
                  example: "barbeiro@trato.com"
      responses:
        "200":
          description: Email de recuperação enviado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email de recuperação enviado"

  # =================== APPOINTMENTS ===================
  /appointments/{id}/schedule-notifications:
    post:
      tags: [Appointments]
      summary: Agendar notificações
      description: |
        Agenda notificações automáticas para um agendamento:
        - 24h antes
        - 1h antes  
        - 15min antes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                channel:
                  type: string
                  enum: [whatsapp, sms, email, push]
                  default: whatsapp
                  example: "whatsapp"
      responses:
        "200":
          description: Notificações agendadas com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Notificações agendadas com sucesso"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "400":
          $ref: "#/components/responses/BadRequest"

  # =================== DASHBOARD ===================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token do Supabase Auth
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        user_metadata:
          type: object
        app_metadata:
          type: object
        created_at:
          type: string
          format: date-time
    Session:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: "bearer"
        expires_in:
          type: integer
          example: 3600
        refresh_token:
          type: string
        user:
          $ref: "#/components/schemas/User"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Mensagem de erro"
        code:
          type: string
          example: "VALIDATION_ERROR"
        details:
          type: object
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Dados inválidos"
            code: "VALIDATION_ERROR"
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Token inválido ou expirado"
            code: "UNAUTHORIZED"
...